kbs/attestation: support attestation service to generate Challenge
Related to #162 .1  This commit allows the Challenge step in RCAR
handshake to be generated due to backend attestation service. For
typical attestation services, a nonce will be generated.

This commit also covers IBM SE + CoCoAS case. In this case, CoCoAS must
be accessed to get a challenge which is specificly used by IBM SE.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>